Extract entire home page to new UserHomePage component
